my-part2 -- finished

+++ src/main/java/com/ssa/dao/DescProjetRepository.java
@@ -3,6 +3,6 @@ package com.ssa.dao;
 import com.ssa.entity.DescProjet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DescProjetRepository extends JpaRepository<DescProjet, Integer> {
+public interface DescProjetRepository extends JpaRepository<DescProjet, Long> {
 
 }
