+++ src/main/java/com/ssa/dao/FicheProjetRepository.java
@@ -3,5 +3,5 @@ package com.ssa.dao;
 import com.ssa.entity.FicheProjet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FicheProjetRepository extends JpaRepository<FicheProjet, Integer > {
+public interface FicheProjetRepository extends JpaRepository<FicheProjet, Long > {
 }
