+++ src/main/java/com/ssa/dao/PJointesRepository.java
@@ -3,5 +3,5 @@ package com.ssa.dao;
 import com.ssa.entity.PJointes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PJointesRepository extends JpaRepository<PJointes, Integer > {
+public interface PJointesRepository extends JpaRepository<PJointes, Long > {
 }
