+++ src/main/java/com/ssa/dao/SuiviChantierRepository.java
@@ -3,5 +3,5 @@ package com.ssa.dao;
 import com.ssa.entity.SuiviChantier;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SuiviChantierRepository extends JpaRepository<SuiviChantier, Integer > {
+public interface SuiviChantierRepository extends JpaRepository<SuiviChantier, Long > {
 }
